cmd/trace-agent: move logic to pkg/trace/agent
This change moves the agent initialization logic into the
`pkg/trace/agent` package, making the `cmd/trace-agent` package slim
with program-starting logic only.
