+++ pkg/trace/agent/log.go
@@ -1,4 +1,4 @@
-package main
+package agent
 
 import (
 	"fmt"
