+++ pkg/trace/agent/sampler.go
@@ -1,4 +1,4 @@
-package main
+package agent
 
 import (
 	"fmt"
@@ -8,7 +8,6 @@ import (
 
 	log "github.com/cihub/seelog"
 
-	"github.com/DataDog/datadog-agent/pkg/trace/agent"
 	"github.com/DataDog/datadog-agent/pkg/trace/config"
 	"github.com/DataDog/datadog-agent/pkg/trace/info"
 	"github.com/DataDog/datadog-agent/pkg/trace/sampler"
@@ -64,7 +63,7 @@ func (s *Sampler) Run() {
 }
 
 // Add samples a trace and returns true if trace was sampled (should be kept), false otherwise
-func (s *Sampler) Add(t agent.ProcessedTrace) (sampled bool, rate float64) {
+func (s *Sampler) Add(t ProcessedTrace) (sampled bool, rate float64) {
 	atomic.AddUint64(&s.totalTraceCount, 1)
 	sampled, rate = s.engine.Sample(t.Trace, t.Root, t.Env)
 	if sampled {
