+++ pkg/trace/agent/service.go
@@ -1,4 +1,4 @@
-package main
+package agent
 
 import (
 	"sync"
